use an unsigned_64 for DiffContextSize and add saturated add/subtract

+++ pkg/config/app_config.go
@@ -457,7 +457,7 @@ type AppState struct {
 
 	HideCommandLog             bool
 	IgnoreWhitespaceInDiffView bool
-	DiffContextSize            int
+	DiffContextSize            uint64
 	RenameSimilarityThreshold  int
 	LocalBranchSortOrder       string
 	RemoteBranchSortOrder      string
